Moved Localization utils and constants to separate files
